# api/cart/controller.go
package cart

import (
	"github.com/gin-gonic/gin"
	"net/http"
	"shopping/domain/cart"
	"shopping/utils/api_helper"
)

type Controller struct {
	cartService *cart.Service
}

func NewCartController(service *cart.Service) *Controller {
	return &Controller{
		cartService: service,
	}
}

// AddItem godoc
// @Summary 添加Item
// @Tags Cart
// @Accept json
// @Produce json
// @Param Authorization header string true "Authentication header"
// @Success 200 {object} api_helper.Response
// @Failure 400 {object} api_helper.ErrResponse
// @Router /cart/item [post]
func (c *Controller) AddItem(g *gin.Context) {
	var req ItemCartRequest
	if err := g.ShouldBind(&req); err != nil {
		api_helper.HandlerError(g, err)
		return
	}
	userId := api_helper.GetUserId(g)
	err := c.cartService.AddItem(userId, req.SKU, req.Count)
	if err != nil {
		api_helper.HandlerError(g, err)
		return
	}
	g.JSON(
		http.StatusCreated, CreateCategoryResponse{
			Message: "Item added to cart",
		})
}

// UpdateItem godoc
// @Summary 更新Item
// @Tags Cart
// @Accept json
// @Produce json
// @Param Authorization header string true "Authentication header"
// @Success 200 {object} api_helper.Response
// @Failure 400 {object} api_helper.ErrResponse
// @Router /cart/item [patch]
func (c *Controller) UpdateItem(g *gin.Context) {
	var req ItemCartRequest
	if err := g.ShouldBind(&req); err != nil {
		api_helper.HandlerError(g, err)
		return
	}
	userId := api_helper.GetUserId(g)
	err := c.cartService.UpdateItem(userId, req.SKU, req.Count)
	if err != nil {
		api_helper.HandlerError(g, err)
		return
	}
	g.JSON(
		http.StatusCreated, CreateCategoryResponse{
			Message: "Updated",
		})
}

// GetCart godoc
// @Summary 获得购物车商品列表
// @Tags Cart
// @Accept json
// @Produce json
// @Param Authorization header string true "Authentication header"
// @Success 200 {object} cart.Item
// @Failure 400 {object} api_helper.ErrResponse
// @Router /cart [get]
func (c *Controller) GetCart(g *gin.Context) {
	userId := api_helper.GetUserId(g)
	result, err := c.cartService.GetCartItems(userId)
	if err != nil {
		api_helper.HandlerError(g, err)
		return
	}
	g.JSON(200, result)
}